+++ packages/sdk/fastly/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## [0.2.6](https://github.com/launchdarkly/js-core/compare/fastly-server-sdk-v0.2.5...fastly-server-sdk-v0.2.6) (2026-02-25)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @launchdarkly/js-server-sdk-common bumped from 2.17.2 to 2.18.0
+
 ## [0.2.5](https://github.com/launchdarkly/js-core/compare/fastly-server-sdk-v0.2.4...fastly-server-sdk-v0.2.5) (2026-02-03)