+++ packages/shared/sdk-server/CHANGELOG.md
@@ -8,6 +8,20 @@ All notable changes to `@launchdarkly/js-server-sdk-common` will be documented i
   * dependencies
     * @launchdarkly/js-sdk-common bumped from 2.3.0 to 2.3.1
 
+## [2.18.0](https://github.com/launchdarkly/js-core/compare/js-server-sdk-common-v2.17.2...js-server-sdk-common-v2.18.0) (2026-02-25)
+
+
+### Features
+
+* Refactor FDV2 protocol handling. ([4570089](https://github.com/launchdarkly/js-core/commit/4570089cd478cc5811a9a1c207231a96fdb5b39a))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @launchdarkly/js-sdk-common bumped from 2.20.0 to 2.21.0
+
 ## [2.17.2](https://github.com/launchdarkly/js-core/compare/js-server-sdk-common-v2.17.1...js-server-sdk-common-v2.17.2) (2026-02-03)