+++ packages/store/node-server-sdk-dynamodb/CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## [6.2.19](https://github.com/launchdarkly/js-core/compare/node-server-sdk-dynamodb-v6.2.18...node-server-sdk-dynamodb-v6.2.19) (2026-02-25)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * devDependencies
+    * @launchdarkly/node-server-sdk bumped from 9.10.6 to 9.10.7
+  * peerDependencies
+    * @launchdarkly/node-server-sdk bumped from >=9.4.3 to >=9.10.7
+
 ## [6.2.18](https://github.com/launchdarkly/js-core/compare/node-server-sdk-dynamodb-v6.2.17...node-server-sdk-dynamodb-v6.2.18) (2026-02-03)